fix 6x gpload to support column without quotes
judge if a into_column name is quoted, if not, add double quotations to it.
Pyyaml cannot check if a string is quoted or not, so we add double quotations to all column names. This will support writting into column like 'Col_name', "Col_name" and Col_name. '"Col_name"' is also supported.